requirements.txt
alembic==1.9.1
autopep8==2.0.1
blinker==1.5
cachelib==0.10.2
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
Deprecated==1.2.13
diff-match-patch==20200713
dnspython==2.2.1
dominate==2.7.0
email-validator==1.3.0
ez-setup==0.9
fernet==1.0.1
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Dance==6.2.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.1
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
gitdb==4.0.10
GitPython==3.1.30
google-api-core==2.11.0
google-api-python-client==2.71.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.57.1
greenlet==2.0.2
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
mysql-connector-python==8.0.32
oauthlib==3.2.2
packaging==22.0
pip-autoremove==0.10.0
protobuf==3.20.3
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyGithub==1.58.1
PyJWT==2.6.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.46
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.13
URLObject==2.4.3
uWSGI==2.0.21
visitor==0.1.3
Werkzeug==2.2.2
wrapt==1.15.0
WTForms==3.0.1
zipp==3.11.0
